Title: 알파벳과 사칙연산, Time: 1221ms, Memory: 176MB, Status: Memory Limit Exceed - CodeTree

+++ 241009/알파벳과 사칙연산/calculations-with-alphabet.java
@@ -9,7 +9,7 @@ public class Main {
     static int res[];
     static int arr[] ={1,2,3,4};
     static int n;
-    static long maxVal = Integer.MIN_VALUE;
+    static long maxVal = Long.MIN_VALUE;
     static String str[];
 
     public static void main(String[] args) throws IOException {
